+++ mayan/apps/file_metadata/views.py
@@ -108,8 +108,12 @@ class DocumentFileSubmitView(MultipleObjectConfirmActionView):
     object_permission = permission_file_metadata_submit
     pk_url_kwarg = 'document_file_id'
     source_queryset = DocumentFile.valid.all()
-    success_message_singular = '%(count)d document file submitted to the file metadata queue.'
-    success_message_plural = '%(count)d documents files submitted to the file metadata queue.'
+    success_message_singular = _(
+        '%(count)d document file submitted to the file metadata queue.'
+    )
+    success_message_plural = _(
+        '%(count)d documents files submitted to the file metadata queue.'
+    )
 
     def get_extra_context(self):
         queryset = self.object_list